rename(TableGroupService, TableService): 패키지 이동
