+++ src/main/java/kitchenpos/table/ui/TableRestController.java
@@ -2,7 +2,7 @@ package kitchenpos.table.ui;
 
 import java.net.URI;
 import java.util.List;
-import kitchenpos.application.TableService;
+import kitchenpos.table.application.TableService;
 import kitchenpos.table.dto.OrderTableRequest;
 import kitchenpos.table.dto.OrderTableResponse;
 import org.springframework.http.ResponseEntity;
